Make the frequency function more flexible
Moved all variables for the frequency function into actual variables. The hook is now ready to be altered to use a dynamic distance instead of a fixed one.

The general exponential function stayed the same, but the equation the get the necessary factors gets solved by the code itself and not manually like before.